Estable agregado modular de productos y bodegas
 Changes to be committed:
	modified:   src/main/java/com/karma/gema/controller/StoresController.java
	modified:   src/main/java/com/karma/gema/model/Company.java
	modified:   src/main/java/com/karma/gema/model/Product.java
	modified:   src/main/java/com/karma/gema/model/Stores.java
	modified:   src/main/java/com/karma/gema/repositories/StoresRepository.java
	modified:   src/main/java/com/karma/gema/request/ProductRequest.java
	modified:   src/main/java/com/karma/gema/request/StoresRequest.java
	modified:   src/main/java/com/karma/gema/service/ProductServiceImpl.java
	modified:   src/main/java/com/karma/gema/service/StoresService.java
	modified:   src/main/java/com/karma/gema/service/StoresServiceImpl.java

 Changes not staged for commit:
	modified:   target/m2e-wtp/web-resources/META-INF/maven/GeMA/GeMA/pom.properties

+++ src/main/java/com/karma/gema/request/StoresRequest.java
@@ -6,7 +6,7 @@ import com.karma.gema.model.Wharehouse;
 public class StoresRequest {
 	private Wharehouse wharehouse;
 	private Product product;
-	private Long amount;
+	private double amount;
 
 	public Wharehouse getWharehouse() {
 		return wharehouse;
@@ -16,7 +16,7 @@ public class StoresRequest {
 		return product;
 	}
 
-	public Long getAmount() {
+	public double getAmount() {
 		return amount;
 	}
 
@@ -28,7 +28,7 @@ public class StoresRequest {
 		this.product = product;
 	}
 
-	public void setAmount(Long amount) {
+	public void setAmount(double amount) {
 		this.amount = amount;
 	}
